pyproject.toml
[project]
name = "mlpops"
version = "0.1.0"
description = "Machine Learning People Operations"
authors = [
    { name = "Yilmaz Mustafa", email = "ymus@tuta.io" }
]
readme = "README.md"
requires-python = ">=3.11,<3.12" 
# limit to 3.11 for now because of dvc

dependencies = [    
    "fastapi>=0.100.0",
    "uvicorn==0.15.0",
    "streamlit==1.32.0",
    "pandas==2.1.3",
    "numpy==1.26.4",
    "scikit-learn>=1.6.1",  
    "networkx==3.2",
    "nltk>=3.9.1",
    "dvc[s3]==3.33.3",
    "fsspec==2023.10.0",
    "imbalanced-learn==0.13.0",
    "pyyaml==6.0.1",
    "python-dotenv==1.0.1",
    "mlflow==2.8.1",
    "prometheus-client==0.18.0",
    "joblib==1.3.2",
    "pydantic>=2.0.0,<3.0.0",
    "crawl4ai>=0.6.0",
    "litellm>=1.0.0",
    "tqdm>=4.66.1",
    "groq>=0.4.0",
    "playwright>=1.40.0",
    "playwright-stealth==1.0.6",
    "instructor>=0.1.0"
]

[project.scripts]
api = "api:main"
web = "web:cli_main"
train = "pipeline:main"

[build-system]
requires = [
    "setuptools>=68.0.0",
    "wheel>=0.42.0",
    "meson-python>=0.18.0"
]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
